refactor(cost): modify cost fields and adding cost data
