+++ src/main/java/com/e2i/wemeet/domain/cost/Cost.java
@@ -1,6 +1,5 @@
-package com.e2i.wemeet.domain.credit;
+package com.e2i.wemeet.domain.cost;
 
-import com.e2i.wemeet.domain.base.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -13,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Cost extends BaseTimeEntity {
+public class Cost {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
