--- src/main/java/com/e2i/wemeet/domain/credit/Earn.java
@@ -1,14 +1,0 @@
-package com.e2i.wemeet.domain.credit;
-
-import lombok.Getter;
-
-@Getter
-public enum Earn {
-    PAYMENT,
-    EVENT,
-    ADVERTISEMENT;
-
-    public static String getTypeName() {
-        return "EARN";
-    }
-}
