+++ src/main/resources/db/migration/V2308030035__create_cost_history.sql
@@ -5,7 +5,6 @@ CREATE TABLE IF NOT EXISTS `cost_history`
     `cost_value`         tinyint     NOT NULL,
     `member_id`          bigint     NOT NULL,
     `create_at`          datetime(6) NOT NULL,
-    `modified_at`        datetime(6) NOT NULL,
     PRIMARY KEY (`cost_history_id`),
     FOREIGN KEY (`member_id`) REFERENCES `member` (`member_id`)
 )
